% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_Admin_User_Table.R
\name{DT_Admin_User_Table}
\alias{DT_Admin_User_Table}
\title{Creates the User Table}
\usage{
DT_Admin_User_Table(user_table)
}
\arguments{
\item{user_table}{The inventory table to generate the report.}
}
\value{
usr_DT A User Table Data Table.
}
\description{
This function creates a User Table Data Table.
}
\details{
Requires: DT, data.table
}
