% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Goals_Update.R
\name{User_Goals_Update}
\alias{User_Goals_Update}
\title{Updates User Goals in Database}
\usage{
User_Goals_Update(u_user_goals, changed_tbl)
}
\arguments{
\item{u_user_goals}{The User Goal table with updates.}

\item{changed_tbl}{The changed fields tables for User Goals.}
}
\description{
This function Updates User Goals in Database.
}
\details{
Requires: dplyr
}
