% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Record_Delete.R
\name{User_Record_Delete}
\alias{User_Record_Delete}
\title{Deletes User Record from Database}
\usage{
User_Record_Delete(d_user_record)
}
\arguments{
\item{d_user_record}{The user record to update in the database.}
}
\description{
This function Deletes User Record from Database.
}
